hotfix

+++ src/main/java/com/agroenvios/clientes/primary/dto/pago/ItemPagoDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ItemPagoDto {
     private String nombre;
-    private int cantidad;
+    private double cantidad;
     private double precio;
 }
